Rename pulsar function api from V1 to V2
- rename the V1 to V2
- change the way how the rest api was organized to allow it merged back to pulsar webservice and cli
